ONNX import main switch refactor checkpoint
