+++ include/eddl/serialization/onnx/layers/conv/upsampling_onnx.h
@@ -8,6 +8,18 @@
  * ONNX EXPORT
  */
 
+// OPSET: 9 (Op deprecated in ONNX)
+Layer* build_upsampling_layer(onnx::NodeProto *node,
+                              map<string, vector<float>> &map_init_values,
+                              map<string, vector<int>> &map_init_dims,
+                              map<string, Layer *> &output_node_map,
+                              int dev,
+                              int mem);
+
+/*
+ * ONNX EXPORT
+ */
+
 // OPSET: 9
 void build_upsample_node(LUpSampling *layer, onnx::GraphProto *graph);
