+++ include/eddl/serialization/onnx/layers/core/activation_onnx.h
@@ -5,6 +5,88 @@
 #include "eddl/layers/core/layer_core.h"
 
 /*
+ * ONNX IMPORT
+ */
+
+// OPSET: 14, 13, 6
+Layer* build_relu_layer(onnx::NodeProto *node,
+                        map<string, Layer *> &output_node_map,
+                        int dev,
+                        int mem);
+
+// OPSET: 13, 6
+Layer* build_sigmoid_layer(onnx::NodeProto *node,
+                           map<string, Layer *> &output_node_map,
+                           int dev,
+                           int mem);
+
+// OPSET: 6
+Layer* build_hard_sigmoid_layer(onnx::NodeProto *node,
+                               map<string, Layer *> &output_node_map,
+                               int dev,
+                               int mem);
+
+// OPSET: 13, 6
+Layer* build_tanh_layer(onnx::NodeProto *node,
+                        map<string, Layer *> &output_node_map,
+                        int dev,
+                        int mem);
+
+// Not in ONNX: Custom operator
+Layer* build_exponential_layer(onnx::NodeProto *node,
+                               map<string, Layer *> &output_node_map,
+                               int dev,
+                               int mem);
+
+// Not in ONNX: Custom operator
+Layer* build_linear_layer(onnx::NodeProto *node,
+                          map<string, Layer *> &output_node_map,
+                          int dev,
+                          int mem);
+
+// OPSET: 6
+Layer* build_leaky_relu_layer(onnx::NodeProto *node,
+                              map<string, Layer *> &output_node_map,
+                              int dev,
+                              int mem);
+
+// OPSET: 10
+Layer* build_thresholded_relu_layer(onnx::NodeProto *node,
+                                    map<string, Layer *> &output_node_map,
+                                    int dev,
+                                    int mem);
+
+// OPSET: 6
+Layer* build_elu_layer(onnx::NodeProto *node,
+                       map<string, Layer *> &output_node_map,
+                       int dev,
+                       int mem);
+
+// OPSET: 6
+Layer* build_selu_layer(onnx::NodeProto *node,
+                       map<string, Layer *> &output_node_map,
+                       int dev,
+                       int mem);
+
+// OPSET: 1
+Layer* build_softsign_layer(onnx::NodeProto *node,
+                            map<string, Layer *> &output_node_map,
+                            int dev,
+                            int mem);
+
+// OPSET: 1
+Layer* build_softplus_layer(onnx::NodeProto *node,
+                            map<string, Layer *> &output_node_map,
+                            int dev,
+                            int mem);
+
+// OPSET: 11, 1
+Layer* build_softmax_layer(onnx::NodeProto *node,
+                           map<string, Layer *> &output_node_map,
+                           int dev,
+                           int mem);
+
+/*
  * ONNX EXPORT
  */
