+++ include/eddl/serialization/onnx/layers/core/dense_onnx.h
@@ -3,6 +3,20 @@
 #define EDDL_DENSE_ONNX_H
 #include "eddl/serialization/onnx/onnx.pb.h"
 #include "eddl/layers/core/layer_core.h"
+#include "eddl/serialization/onnx/utils_onnx.h"
+
+/*
+ * ONNX EXPORT
+ */
+
+// OPSET: 13, 11
+Layer* build_dense_layer(onnx::NodeProto *node,
+                         map<string, vector<float>> &map_init_values,
+                         map<string, vector<int>> &map_init_dims,
+                         map<string, Layer *> &output_node_map,
+                         LOG_LEVEL log_level,
+                         int dev,
+                         int mem);
 
 /*
  * ONNX EXPORT
