+++ src/serialization/onnx/net/layers/conv/conv_onnx.cpp
@@ -1,6 +1,136 @@
 #if defined(cPROTO)
 #include "eddl/serialization/onnx/layers/conv/conv_onnx.h"
+#include "eddl/serialization/onnx/utils_onnx.h"
 
+// ONNX import
+Layer* build_conv_layer(onnx::NodeProto *node,
+                        map<string, vector<float>> &map_init_values,
+                        map<string, vector<int>> &map_init_dims,
+                        map<string, Layer *> &output_node_map,
+                        int dev,
+                        int mem)
+{
+  int filters;
+  vector<int> kernel_shape;
+  vector<int> strides;
+  vector<int> pads = {};
+  string auto_pad_option = "custom";
+  vector<float> *bias;
+  bool use_bias = node->input_size() > 2;
+  bool conv1d = false;
+  int groups = 1;
+  vector<int> dilation_rate = {1, 1};
+
+  for (int j = 0; j < node->attribute_size(); j++)
+  { // Set the attributes
+    onnx::AttributeProto attribute = node->attribute(j);
+    string attr_name = attribute.name();
+    if (!attr_name.compare("auto_pad"))
+    {
+      if (!attribute.s().compare("NOTSET"))
+        auto_pad_option = "custom";
+      else if (!attribute.s().compare("VALID"))
+        auto_pad_option = "valid";
+      else if (!attribute.s().compare("SAME_UPPER"))
+        auto_pad_option = "same";
+    }
+    //else if (!attr_name.compare("dilations")) { It isn't implemented in eddl
+    //}
+    //else if (!attr_name.compare("group")) { It isn't implemented in eddl
+    //}
+    else if (!attr_name.compare("kernel_shape"))
+    {
+      for (int h = 0; h < attribute.ints_size(); h++)
+      {
+        kernel_shape.push_back(attribute.ints(h));
+      }
+      if (attribute.ints_size() == 1)
+      { // If is conv1D, we make the equivalent in conv2D
+        conv1d = true;
+      }
+    }
+    else if (!attr_name.compare("pads"))
+    {
+      for (int h = 0; h < attribute.ints_size(); h++)
+      {
+        pads.push_back(attribute.ints(h));
+      }
+      if (attribute.ints_size() == 4)
+        swap(pads[1], pads[2]);
+    }
+    else if (!attr_name.compare("strides"))
+    {
+      for (int h = 0; h < attribute.ints_size(); h++)
+      {
+        strides.push_back(attribute.ints(h));
+      }
+      if (attribute.ints_size() == 1)
+      { // If is conv1D, we make the equivalent in conv2D
+        conv1d = true;
+      }
+    }
+  }
+
+  string parent_name = node->input(0); // Get parent
+  Layer *parent = output_node_map[parent_name];
+  vector<int> parent_shape = parent->output->shape;
+
+  string weights_name = node->input(1); // Get weights and dims
+  vector<float> *weights = &(map_init_values[weights_name]);
+  vector<int> dims = map_init_dims[weights_name];
+
+  if (parent_shape.size() == 3)
+  {
+    conv1d = true;
+  }
+
+  if (conv1d)
+  {
+    strides.push_back(1);
+    kernel_shape.push_back(1);
+    dims.push_back(1);
+    pads.push_back(0);
+    pads.push_back(0);
+  }
+
+  filters = dims[0];
+  string name = node->name();
+  ConvolDescriptor *cd = new ConvolDescriptor(filters, 
+                                              kernel_shape, 
+                                              strides, 
+                                              auto_pad_option,
+                                              pads, 
+                                              groups, 
+                                              dilation_rate, 
+                                              use_bias, 
+                                              mem);
+
+  Layer *actual_layer;
+  if (conv1d)
+    actual_layer = new LConv1D(parent, cd, name, dev, mem);
+  else
+    actual_layer = new LConv(parent, cd, name, dev, mem);
+
+  if (use_bias)
+  {
+    string bias_name = node->input(2);
+    bias = &(map_init_values[bias_name]);
+    vector<int> bias_shape;
+    bias_shape.push_back(bias->size());
+    Tensor *bias_tensor = new Tensor(bias_shape, nullptr, dev);
+    COPY_FROM_VECTOR_PTR_TO_TENSOR(bias, bias_tensor);
+    Tensor::copy(bias_tensor, cd->bias);
+    delete bias_tensor;
+  }
+  Tensor *weights_tensor = new Tensor(dims, nullptr, dev);
+  COPY_FROM_VECTOR_PTR_TO_TENSOR(weights, weights_tensor);
+  Tensor::copy(weights_tensor, cd->K);
+  delete weights_tensor;
+
+  return actual_layer;
+}
+
+// ONNX export
 void build_conv_node(LConv *layer, onnx::GraphProto *graph, bool gradients)
 {
   // Add an empty node to the graph
