+++ src/serialization/onnx/net/utils_onnx.cpp
@@ -20,6 +20,115 @@ std::vector<int> vf2vi(const std::vector<float> &vf)
   return vi;
 }
 
+vector<float> parseTensorValues(onnx::TensorProto t)
+{
+  int data_type = t.data_type(); // Only works for non raw data for now
+  vector<float> values;
+  switch (data_type)
+  {
+  case onnx::TensorProto::FLOAT:
+    if (t.has_raw_data())
+    {
+      TryConvertingTensorRawValues(t, values);
+    }
+    else
+    {
+      for (int i = 0; i < t.float_data_size(); i++)
+      {
+        values.push_back(t.float_data(i));
+      }
+    }
+    break;
+  case onnx::TensorProto::UINT8:
+    for (int i = 0; i < t.int32_data_size(); i++)
+    {
+      values.push_back(t.int32_data(i));
+    }
+    break;
+  case onnx::TensorProto::INT8:
+    for (int i = 0; i < t.int32_data_size(); i++)
+    {
+      values.push_back(t.int32_data(i));
+    }
+    break;
+  case onnx::TensorProto::UINT16:
+    for (int i = 0; i < t.int32_data_size(); i++)
+    {
+      values.push_back(t.int32_data(i));
+    }
+    break;
+  case onnx::TensorProto::INT16:
+    for (int i = 0; i < t.int32_data_size(); i++)
+    {
+      values.push_back(t.int32_data(i));
+    }
+    break;
+  case onnx::TensorProto::INT32:
+    for (int i = 0; i < t.int32_data_size(); i++)
+    {
+      values.push_back(t.int32_data(i));
+    }
+    break;
+  case onnx::TensorProto::INT64:
+    if (t.has_raw_data())
+    {
+      vector<int64_t> aux_values; // Vector to read the int64 values
+      TryConvertingTensorRawValues(t, aux_values);
+      for (float i : aux_values) // Cast to float
+        values.push_back(i);
+    }
+    else
+    {
+      for (int i = 0; i < t.int64_data_size(); i++)
+      {
+        values.push_back(t.int64_data(i));
+      }
+    }
+    break;
+  case onnx::TensorProto::BOOL:
+    for (int i = 0; i < t.int32_data_size(); i++)
+    {
+      values.push_back(t.int32_data(i));
+    }
+    break;
+  case onnx::TensorProto::FLOAT16:
+    break;
+  case onnx::TensorProto::DOUBLE:
+    for (int i = 0; i < t.double_data_size(); i++)
+    {
+      values.push_back(t.double_data(i));
+    }
+    break;
+  case onnx::TensorProto::UINT32:
+    for (int i = 0; i < t.uint64_data_size(); i++)
+    {
+      values.push_back(t.uint64_data(i));
+    }
+    break;
+  case onnx::TensorProto::UINT64:
+    for (int i = 0; i < t.uint64_data_size(); i++)
+    {
+      values.push_back(t.uint64_data(i));
+    }
+    break;
+  // TODO
+  //case onnx::TensorProto::STRING:
+  //  break;
+  //case onnx::TensorProto::UNDEFINED:
+  //  break;
+  //case onnx::TensorProto::COMPLEX64:
+  //  break;
+  //case onnx::TensorProto::COMPLEX128:
+  //  break;
+  //case onnx::TensorProto::BFLOAT16:
+  //  break;
+  default:
+    cerr << "Vector type not recognized" << endl;
+    break;
+  }
+  return values;
+}
+
 void sync_snets_with_orig(Net *net, bool acc_gradients)
 {
   if (net->snets[0]->dev != DEV_CPU)
